Move TextShortcut to input.rs and rename Shortcut to StandardShortcut
